fix uncaught asm minor change leading to bug
https://github.com/twmb/murmur3/pull/4#discussion_r64139621 was
addressed partially, but the JNE at the end was not changed.

Prior, a CMP was before the JNE. Now it's LEA. Coincidentally, the
LEA would result in zero if R12 were zero.

This would've only happened if:
- The input seed were zero and the first 4 bytes were zero
- The input seed were zero and the LEA of prior mutations resulted in
  zero

The second case is unlikely. The first case was not caught because the
Incremental test was changed to have pure random input.

Quick checks didn't catch this, I'm guessing, because there is no
standard byte generation, so each byte was generated randomly -
(1/256)^4 chance of the first word being zero, and 1/(math.MaxUint32) of
the seed being zero.

The prior commit, which checks all zeros on the first iteration of
Incremental, catches this bug.