Add Go compiler directive 'noescape' to assembly implementations
Without this directive Go considers that `data []byte` escapes from the
function, and forces allocation on the heap even for values which can
perfectly stay on the stack.

This removes extra allocation when value passed in is not on the heap
already.

See also: https://github.com/golang/go/issues/4099