update to fix 1.14 failures
Unaligned reads cause panics on 1.14, which I guess makes sense on some
architectures, but how we were doing it on intel should be safe. But
this is a general package, so.

Switches to binary.LittleEndian to read four bytes at a time at the cost
of a 15%ish performance hit.

The assembly for amd64 still works because the unaligned loads being
done work on Intel at the cost of basically double the cpu. I don't
expect more input to be unaligned, so this is likely an ok tradeoff.

Lastly, switches to math/bits for rotates because why not.