remove reflect, hash on strings
This eliminates one ADDQ from the hot loop, which speeds things up a
tiny bit:

name             old time/op    new time/op    delta
32Sizes/8192-12    2.20µs ± 2%    2.19µs ± 3%   ~     (p=0.302 n=9+19)

name             old speed      new speed      delta
32Sizes/8192-12  3.72GB/s ± 2%  3.75GB/s ± 3%   ~     (p=0.308 n=9+19)

Unfortunately, the remaining logic cannot be removed. Here is the
difference between this compilation and spaolacci's compilation (which
is on the old unsafe code):

Mine:
	0x0021 00033 (murmur32_gen.go:36)	MOVL	(SI), DI
	0x0023 00035 (murmur32_gen.go:37)	ADDQ	$-4, DX
	0x0027 00039 (murmur32_gen.go:37)	MOVQ	DX, R8
	0x002a 00042 (murmur32_gen.go:37)	NEGQ	DX
	0x002d 00045 (murmur32_gen.go:37)	SARQ	$63, DX
	0x0031 00049 (murmur32_gen.go:37)	ANDQ	$4, DX
	0x0035 00053 (murmur32_gen.go:37)	ADDQ	DX, SI

Spaolacci's:
	0x0035 00053 (murmur32.go:124)	MOVQ	DX, R8
	0x0038 00056 (murmur32.go:123)	ADDQ	$4, R8
	0x003c 00060 (murmur32.go:124)	MOVL	(DX), R9

So, the old code just unconditionally adds four and loads the word.

The new code loads the word and then does logic to _only_ add four _if_
we are not past the end already. This is a smart compiler rewrite that
avoids a jump, but unfortuantely this jump avoidance adds 4 instructions
in a loop that we don't need. At this point, remaining performance is a
compiler issue.

Side note: I am happy to see that the little endian load is recognized
as just a pure load on my arch, but that also means that the performance
of these loops is likely different on big endian by 40%. Oh well!