all: switch to reflect for safer slice->string conversions
Go issue 19367 (https://github.com/golang/go/issues/19367)
and a linked issue at the bottom indicate that the following trick

  *(*string)(unsafe.Pointer(&slice))

is not theoretically safe. While I doubt that the Go people will ever
actually break that code since it would break a loooot of other code, we
may as well convert back to safe code. The prior commit removed reflect
because we were converting from a string to a slice; this re-introduces
it because converting from a slice to a string adds only 0.4ns of
overhead but adds following the rules of Go to the letter.

After this commit, this library effectively follows all rules of
unsafety (and is likely one of the few murmur3 libraries that does so).