Make strslice more portable
TinyGo uses a slightly different layout of strings and slices (which is
only relevant on AVR): it uses uintptr instead of int for the Len and
Cap fields. The reflect.StringHeader and reflect.SliceHeader were never
portable to begin with, but this change makes the code slightly more
portable to include TinyGo.

I've checked the performance with `go test -bench=Strslice` but the
difference appears to be well within the noise.